More refactors for buffer and core ABR functionality seams.
