+++ src/fp/toMerged.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const toMerged = (...os) => Object.assign({}, ...os);
+export { toMerged };
+export default toMerged;
